requirements.txt
rdflib==6.3.2
chromadb==0.4.18
numpy>=1.21.0
openai>=1.3.7
python-dotenv==1.0.0
PyYAML==6.0.1
websockets>=12.0
anyio>=4.6 